etl/requirements.txt
boto3
psycopg2-binary
pandas
pyarrow
python-dotenv